SB800: Port to 64bit
Reviewed-on: http://review.coreboot.org/10582
Tested-by: build bot (Jenkins)
Tested-by: Raptor Engineering Automated Test Stand <noreply@raptorengineeringinc.com>

+++ src/southbridge/amd/cimx/sb800/spi.c
@@ -35,7 +35,7 @@ static int bus_claimed = 0;
 
 #define AMD_SB_SPI_TX_LEN	8
 
-static u32 spibar;
+static uintptr_t spibar;
 
 static void reset_internal_fifo_pointer(void)
 {
